requirements.txt
flask==2.2.5
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2 